backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
base58==2.1.1
bitarray==2.9.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
cytoolz==0.12.3
distlib==0.3.8
Django==5.1
django-admin-interface==0.30.1
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-grappelli==4.0.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
env==0.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.7.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.15.4
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.8
inflection==0.5.1
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
lru-dict==1.3.0
multiaddr==0.0.9
multidict==6.0.5
mysqlclient==2.2.4
netaddr==1.3.0
packaging==24.1
parsimonious==0.8.1
pillow==11.0.0
platformdirs==4.2.2
protobuf==3.20.3
pycryptodome==3.20.0
PyJWT==2.9.0
PyMySQL==1.1.1
pypng==0.20220715.0
PyQRCode==1.2.1
pyrsistent==0.20.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
requests==2.32.3
rlp==2.0.1
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
toolz==0.12.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
varint==1.0.2
virtualenv==20.26.3
web3==5.28.0
websockets==9.1
yarl==1.9.4
gunicorn
whitenoise
